Rasterize at least something

+++ include/Utilities.h
@@ -43,8 +43,8 @@ namespace OsmAnd {
         OSMAND_CORE_API double OSMAND_CORE_CALL toRadians(double angle);
         OSMAND_CORE_API double OSMAND_CORE_CALL get31TileNumberX(double longitude);
         OSMAND_CORE_API double OSMAND_CORE_CALL get31TileNumberY(double latitude);
-        OSMAND_CORE_API double OSMAND_CORE_CALL get31LongitudeX(uint32_t x);
-        OSMAND_CORE_API double OSMAND_CORE_CALL get31LatitudeY(uint32_t y);
+        OSMAND_CORE_API double OSMAND_CORE_CALL get31LongitudeX(double x);
+        OSMAND_CORE_API double OSMAND_CORE_CALL get31LatitudeY(double y);
         OSMAND_CORE_API double OSMAND_CORE_CALL getTileNumberX(float zoom, double longitude);
         OSMAND_CORE_API double OSMAND_CORE_CALL getTileNumberY(float zoom, double latitude);
         OSMAND_CORE_API double OSMAND_CORE_CALL checkLatitude(double latitude);
